requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.5.0
billiard==3.6.4.0
boto3==1.24.80
botocore==1.27.80
celer==0.7.2
celery==5.2.7
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
contourpy==1.0.7
cron-descriptor==1.2.35
cryptography==38.0.1
cycler==0.11.0
Cython==0.29.34
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.13
django-allauth==0.51.0
django-celery==3.3.1
django-celery-beat==2.5.0
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-import-export==2.8.0
django-storages==1.13.1
django-timezone-field==5.0
download==0.3.5
et-xmlfile==1.1.0
fonttools==4.39.3
idna==3.4
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
libsvmdata==0.4.1
MarkupPy==1.14
matplotlib==3.7.1
numpy==1.21.6
oauthlib==3.2.1
odfpy==1.4.1
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==23.1
pandas==2.0.1
Pillow==9.1.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
pyzbar==0.1.9
qrcode==6.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
types-cryptography==3.3.23
typing_extensions==4.2.0
tzdata==2023.3
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
xarray==2023.4.2
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
